Bug Fix #1
Fixed a couple of bugs. Drawing the player should have went with the
tiles draw value instead of its coordinate value. (Its late and I'm not
very coherent so that probably makes no sense.) In game.js I just
changed it to a keydown event instead of a keyup (Makes more sense in my
mind.)